Small bug fixes (interferences between some modes and panels, sometimes related to folding)

+++ pyqode/core/panels/encodings.py
@@ -149,11 +149,11 @@ class EncodingPanel(Panel):
             EncodingsContextMenu(parent=editor)
 
     def _reload(self):
+        self.hide()
         self._rm_deco()
         self.editor.setReadOnly(False)
         self.enable_caret_line(True)
         self.editor.file.reload(self.ui.comboBoxEncodings.current_encoding)
-        self.hide()
 
     def _edit_anyway(self):
         self._rm_deco()
@@ -167,7 +167,8 @@ class EncodingPanel(Panel):
             self._deco = None
 
     def cancel(self):
-        self.editor.clear()
+        if self.sender():
+            self.editor.clear()
         self._rm_deco()
         self.enable_caret_line(True)
         self.cancel_requested.emit(self.editor)
